AliExpress_Shopify_Sync/requirements.txt
ShopifyAPI
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
termcolor
openpyxl
